Fixed an error with displaying the current menu button. Fixed an error with editing leader information
Since not all possible screens have buttons associated with them, when
the changeScreen method tried to change the button class, it sometimes
caused an error.
When a leader's information was edited, an error saying the leader
already exists was given, since the name was technically being
resubmitted. It has been changed to only give an error if the name is
changed to another existing leader's name.